main.star: reorganize postsubmit consoles once again
This change reorganizes postsubmit consoles once again to make it easier
for the release rotation to identify issues.

Firstly, this CL drops the console headers. I think it was a nice idea,
but it's mostly visual noise.

Second, this CL adds the concept of known-issue builder types.

Third, this CL separates out known-issue builders into their own
consoles and puts links to the issues in the header.

Fourth, this CL unifies console naming in the console_name function and
creates the notion of "tiers" for grouping consoles on the console list
page (it's ordered by console name, not title).

Fifth, it moves the by-go-commit consoles into their own tier. On the
main console list page it generally just duplicates information but we
still want the links, and we want them to show up above known-issue
consoles.

And lastly, it updates the console titles once again to be more
human-friendly (for example, by replacing hyphens with whitespace).

Change-Id: I560fe9b536fd0fe8fd83bf52e2caa66a6e7a6375
Reviewed-on: https://go-review.googlesource.com/c/build/+/557935
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>