Revert "main.star: reorganize postsubmit consoles once again"
This reverts CL 557935.

Reason for revert: This CL changed the name of the consoles, resulting
in a loss of history. This is probably not good at this point. It also
turns out the ordering of consoles has nothing to do with the name. I
think this is something I already tripped over in the past, but forgot.
We need documentation on this.

Change-Id: Ic2f3d5666bff8ba57d2d0d00ed6066640eb6453a
Reviewed-on: https://go-review.googlesource.com/c/build/+/558436
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>