internal/task: add a workflow for publishing private x/ patches
This adds a workflow to relui which takes a patch from the private
internal gerrit instance for one of the golang.org/x/ repos and sends it
to the public gerrit, waits for it to be submitted, tags the repo, and
emails an announcement message to the various lists.

Requires a minor change to the HTML workflow template to allow for
non-slice textareas.

Updates golang/go#65756

Change-Id: Ica1ec5982545ddd7fff1e71bd33eb3281572017d
Reviewed-on: https://go-review.googlesource.com/c/build/+/559295
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Roland Shoemaker <roland@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>