main.star: match previous environment in linux-arm builder
The coordinator version of this builder is running well and passing,
but my attempts to start up a LUCI version of the same builder so far
are running into frequent stalls and other signs of corruption, and
infrequent successful runs.

Try to zero in on what's causing the difference by first matching the
environment configuration of the previous builder and see if that has
any effect, or if it's something else.

For golang/go#65241.

Change-Id: Id7f00a998056368a42bce8c6e02f01abbdd5455f
Reviewed-on: https://go-review.googlesource.com/c/build/+/559500
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>