main.star: update for linux/arm builder reporting its arch as 32-bit arm
I've realized it's possible to use setarch on an arm64 host and get
it to report its architecture as 32-bit arm. That works, and causes
the Swarming Bot to report its cipd_platform as linux-armv6l, which
should produce a more useful signal for the purpose of coverage for
the linux/arm port.

Update is_fully_supported to say yes for cipd_platform:linux-armv6l.
That is still my intention, it just needs to be done explicitly now.

For golang/go#65241.

Change-Id: I01c94ac6d3df20b49ec35a2aff1bb585eb73a1c6
Reviewed-on: https://go-review.googlesource.com/c/build/+/560464
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Pratt <mpratt@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>