env/linux-x86-bullseye: add zlib1g to dependencies to install
This dependency is required for clang's llvm-symbolizer to decompress
DWARF debug information in the binaries we (and clang) produce. This is
already accounted for indirectly with the debian package manually
included for linux-x86-clang, but in the LUCI world, I chose to add
clang to CIPD so it's easier to update.

Fixes golang/go#65469.

Change-Id: Id58fb754f959c1a3b6da1b7505003697685a3b45
Reviewed-on: https://go-review.googlesource.com/c/build/+/561595
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>