main.star: use gopls-release-branch.0.11 as the x/tools perf baseline
I misunderstood the baseline x/tools version that should be used for
benchmarking as the v0.11.0 tag. The intended baseline is actually the
gopls-release-branch.0.11 branch. Switch to that and change the run_mod
to be clearer. This will also reset the build history (all builds have
failed so far anyway.)

Change-Id: I95f3ed51773b38c3318dbfaaba414fa4d9f1c6bf
Reviewed-on: https://go-review.googlesource.com/c/build/+/562956
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>