main.star fix protobuf luci environment
When I pushed a new version to
golang/third_party/protoc_with_conformance/linux-amd64 I named the
binaries incorrectly and thus the builders were failing. To fix this I
pushed a fixed version with the same tag. Now the builders are failing
due to ambiguity in the tags (two instances with the same tag). Because
I lack the permissions to clean this up, I pushed a third version with
the correctly named binaries and a different version. This change
references this new version.

Change-Id: I7e40247c369f18e4f9ce22474dac1363ec3ef18d
Reviewed-on: https://go-review.googlesource.com/c/build/+/563295
TryBot-Bypass: Lasse Folger <lassefolger@google.com>
Reviewed-by: Michael Stapelberg <stapelberg@google.com>