internal/relui, internal/task: skip perf and known issue builders
When it comes to planning advisory builders during the release process,
as well as determining commit greenness as part of the tagging process,
we want to skip custom builders (e.g., the new perf builder which does
benchmarking) and builders with known issues (e.g., the linux/arm port
builder which is undergoing the turn-up process).

This change implements that behavior in the two places. Log the causes
of these skips, but default to skipping for custom modes that might be
added in the future. We want that to be possible without incurring any
disruption to these workflows.

Fixes golang/go#65589.

Change-Id: I6bcac92e1c2f298f5a748581ae333402a90d39ef
Reviewed-on: https://go-review.googlesource.com/c/build/+/563395
Reviewed-by: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>