main.star: add linux-arm builder to the internal host
A new distinct set of VMs will be used for security
pre-submit and release testing.

We choose not to have a shared pool in the security
environment at this time, so take that into account.

For golang/go#65241.

Change-Id: I9ef47b7cc91b531b54be3dba3fcd35489a554ebf
Reviewed-on: https://go-review.googlesource.com/c/build/+/564835
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>