cmd/gomote: change what is logged during a bootstrap push
The process for adding a bootstrap version of Go on gomote instances
has changed. It no longer places it in a go1.4 directory. It also
places the bootstrap version of go in a directory inside of the
working directory even though the bootstrap version of go is already
installed by CIPD for the instance. This is an intermediary step to
clarify the logging and remove non-functioning logic.

Change-Id: I46f18837b63483e8b9593865c3df8e090252a358
Reviewed-on: https://go-review.googlesource.com/c/build/+/564836
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>