cmd/gerritbot: migrate Cq-Include-Trybot footers
If a Cq-Include-Trybot footer appears in the GitHub PR message,
include it in the footer section of the Gerrit commit message.

This footer controls the LUCI SlowBots that run on a CL, and has no
effect if it isn't in the footer block, so it is important that we place
it in the footer specifically.

Most GitHub PR users won't add these initially, but reviewers may
request that they add them to add important SlowBots.

Though LUCI requires this be a footer, gerritbot is intentionally more
lenient and allows it anywhere in the message to try to catch simple
mistakes.

Change-Id: Ib31762765aad69e9229501fe3fe3d8f8b6b97eda
Reviewed-on: https://go-review.googlesource.com/c/build/+/564837
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>