relnote: autolink text is code
Autolinks arise from symbols in the release note that are enclosed in
square brackets, like

   See [bytes.Buffer] for details.

The text in square brackets becomes a link to the documentation for
that symbol.

This CL renders that text as code instead of plain text. That is,
before this CL, the above line would be converted to

    See [bytes.Buffer](/pkg/bytes#Buffer) for details.

but this CL adds backticks:

    See [`bytes.Buffer`](/pkg/bytes#Buffer) for details.

For golang/go#64169.

Change-Id: I30f203cf05de535cb8bb200d89912be4c296f34e
Reviewed-on: https://go-review.googlesource.com/c/build/+/577295
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>