cmd/coordinator: rewrite scheme-less URLs and redirects in dev mode
This was created in preparation of removal of dashboard v2, which was
going to use a redirect. However dashboard v2 turned out to be useful
enough to warrant keeping and maintaining it.

Apply this enhancement to the local development serving path since it
is generally useful and already prepared.

For golang/go#65913.

Change-Id: I6006bfa02d512675b63773c22c9ed21d8d5b4ab4
Reviewed-on: https://go-review.googlesource.com/c/build/+/567497
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>