cmd/coordinator/internal/legacydash: refactor package-scoped variables
I was on the of fence whether to take this on so late in the game; it
seems to favor slightly towards doing it anyway. This makes it easier
to see where the variables are used, and will make the future changes
easier to reason about.

For golang/go#65913.

Change-Id: I3c274d2aa7174a9fbd6be91869ce4b1da0dfecaa
Reviewed-on: https://go-review.googlesource.com/c/build/+/567499
Reviewed-by: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>