cmd/relui: allow picking which of the private go projects to build from
Rather than hardcoding using golang/go-private, allow picking from that
and the "go" project on the go-internal host. This paves the way for
migrating to the new repository without needing to rollback in case of
an issue. Once we are fully migrated we'll likely go back to hardcoding
the project.

Change-Id: I380d3934279f7217964bfa5656cba642da11d66e
Reviewed-on: https://go-review.googlesource.com/c/build/+/569457
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>