perf: fix the dashboard landing page
Currently the dashboard landing page is broken because no data for some
of the benchmarks can be found. The reason for this is that we switched
to 16 core machines, so the old -8 benchmarks don't have data in the
last 30 days.

This change switches to using errors.Is for errBenchmarkNotFound. The
reason for the poor error message is because this error gets wrapped,
and we're not searching for it correctly. Using errors.Is let's us check
for the error more robustly. Also, if we can't find data on a particular
benchmark on the default page, I think it's OK to just skip it in
general. It'll be fairly obvious if something is wrong and all the data
is missing.

Change-Id: Iaad6700a6cb35d7ec1b2444b422d35fd55631f41
Reviewed-on: https://go-review.googlesource.com/c/build/+/570238
Reviewed-by: David Chase <drchase@google.com>
Reviewed-by: Michael Pratt <mpratt@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>