relnote: handle symbol links with backticks
Support symbol links in release notes that already are written
as a markdown code element.

For example,

   This release note mentions [`byte.Buffer`].

Previously we handled this properly only if the backticks were omitted.

The parsed markdown for this case is not a single Plain element, which
is what the original code was designed for, but a Plain, then a Code,
then another Plain. It was simpler overall to redesign the algorithm,
first preprocessing the input to isolate the square brackets, then
looking for triples of '[', Plain-or-Code, ']'.

For golang/go#64169.

Change-Id: Ia3036cac2d851efb43c625ccb58831a6fe2c00b6
Reviewed-on: https://go-review.googlesource.com/c/build/+/577260
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>