main.star: make shadow bucket builds publicly readable
We'd like to be able to show the results of a shadow build to anyone,
and they're not expected to have anything private—we'd use something
else for that if needed.

Remove the creator_groups parameter so that its ["mdb/golang-team"]
value is closer to the "# Allow anyone on the Go team to …" comment.
In general, make it clear that these shadow buckets are not intended
for testing anything that's private (e.g., security fix content).

Change-Id: Idd3ab9e6418d9ac5e35c84a49081bb3030dbd757
Reviewed-on: https://go-review.googlesource.com/c/build/+/577855
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Than McIntosh <thanm@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>