cmd/watchflakes: fetch GitHub token from Secret Manager
Unfortunately, the only API that creates a github.Client requires that
credentials for api.github.com are available in $HOME/.netrc. It would
be nicer to be able to pass a custom *http.Client. For now, modify the
global default HTTP client (http.DefaultClient) that the rsc.io/github
package always uses and provide authentication to it directly that way,
without needing to write to a $HOME/.netrc file. This can go away when
the rsc.io/github API provides a cleaner way of doing this.

For golang/go#65353.

Change-Id: Ie2ec3e957f7b2dc48b85ce5475a09d19df71622d
Reviewed-on: https://go-review.googlesource.com/c/build/+/579316
Reviewed-by: Cherry Mui <cherryyz@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>