cmd/watchflakes: ignore failed builds without output properties
It turns out to be possible¹ for a build to fail early enough that it
doesn't get a chance to produce output properties, and not get marked
as an infrastructure failure (even though it should've been). There's
not much choice but to ignore such builds.

¹ For example, https://ci.chromium.org/b/8752293940218446529. It failed
  very early while running bbagent, before golangbuild was even started.

For golang/go#65353.

Change-Id: I14782f41e7a390380a32272b58cda385c91cb881
Reviewed-on: https://go-review.googlesource.com/c/build/+/579299
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>