main.star: filter openbsd/ppc64 from tests on 1.21, add a timeout scale
The port wasn't added until Go 1.22, so no need to test it with Go 1.21.

Also set a timeout scale factor of 2 for now, while the LUCI builder is
running alongside the coordinator builder on the same hardware. This is
fine to adjust later as it becomes more clear what the optimal value is.

For golang/go#63480.
For golang/go#56001.

Change-Id: I707ffe7d15afa6a70d6d8789f959a5835259df3f
Reviewed-on: https://go-review.googlesource.com/c/build/+/585217
Reviewed-by: Cherry Mui <cherryyz@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>