env: add nano and ed to select images
It's currently pretty annoying to do small edits on a gomote when ssh'd
in, since you basically need to edit locally and push. nano is a very
small dependency but would ease a lot of frustration, so add it to the
images.

Change-Id: I4e3747408d9b78eb6d3448e8f124c017e4bdb1ab
Reviewed-on: https://go-review.googlesource.com/c/build/+/587615
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>