perf: make the dashboard charts' X axes based on the real commit history
A long-standing problem with the existing charts is that they always
expand out all data to fit the chart, since each commit is given equal
space on each chart. For benchmarks that go away, or change names, or
are added, this makes comparing two charts side-by-side difficult and
confusing.

The crux of the problem is that the commit list the charts are based on
is the list of commits that actually had benchmark runs for a given
builder.

This change modifies the charts' X axes to be based on the real commit
history of the repository, which is fetched by the server and passed
down to the client.

This also fixes a graphical bug with some of the binary size benchmarks
which have multiple data points for the same commit. Because the
benchmark samples are projected onto a separate history, it'll prevent
the squeezing that happens.

Change-Id: I8b341434b5443fa3bd4386b263dfef659bd509dc
Reviewed-on: https://go-review.googlesource.com/c/build/+/587617
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Pratt <mpratt@google.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>