perf: make clicking on a commit in the chart take you to the unit page
Currently, clicking a commit on the chart takes you to the commit in
gitiles, but this is less useful than seeing the run's details and in
context with runs for nearby commits on the unit page. This CL
implements the latter. An additional benefit of this change is it makes
the unit page more discoverable.

While we're here, fix a bug where non-unit query params aren't passed
through correctly.

Change-Id: I35606efea32e05d60bb5afa580c022bf73240ce3
Reviewed-on: https://go-review.googlesource.com/c/build/+/588155
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Michael Pratt <mpratt@google.com>