perf: filter out commits that don't belong to the history in charts
This change adds a data cleaning step to the charts that removes points
which don't correspond to a commit that appears in the history. The
purpose of this change is primarily to work around a bug in golangbuild
that caused subrepo benchmark results to get tagged with a commit from
the main Go repository. This results in (artistically) broken charts for
subrepo benchmarks.

Change-Id: I4090c392f81f49d08a1aedcb8d77a5fbf9bf7bb7
Reviewed-on: https://go-review.googlesource.com/c/build/+/588136
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Pratt <mpratt@google.com>