internal/releasetargets: regenerate all ports for Go 1.23
The release freeze for Go 1.23 has started, so now is a good
time to regenerate this list. Done with 'go generate'.

The list is the same as for Go 1.22 with the addition of the
new openbsd/riscv64 port.

For golang/go#40561.
For golang/go#55999.

Change-Id: Ib920c3b4337fcd9aa4a765e5622e6c158f829bda
Reviewed-on: https://go-review.googlesource.com/c/build/+/588137
Reviewed-by: Carlos Amedee <carlos@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>