main.star: enable longtest builders on x/benchmarks for sweet
It's already annoying to have to ask for longtest builders to run the
end-to-end test for sweet, but the fact that we also forget to run the
Go 1.22 end-to-end test means we get real breakages like
golang/go#68555.

Resolve this once and for all by enabling longtest in presubmit if any
Go file is touched by a CL in the x/benchmarks repository.

Change-Id: Ia76258edf9bc93b88501cd64755beddd39186f06
Reviewed-on: https://go-review.googlesource.com/c/build/+/600495
Auto-Submit: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>