cmd/retrybuilds: rewrite tool to support LUCI
This change rewrites retrybuilds from the ground up to support LUCI. It
works in a similar way to the old command but eschews some of its
functionality so we can iterate more quickly.

The general structure of the tool is to query buildbucket for builds
based on some predicate and then retrigger them as closely as possbile
to how they were triggered in the first place.

The tool supports several convenience flags for influencing the
predicate (builder ID and build ID) as well as an ability to iterate
over commit ranges (which requires making several queries). Where we're
missing functionality there's also a -predicate flag for directly
modifying the BuildPredicate proto as JSON to allow for full control
over the queries made.

This tool works on all repositories, not just the main Go repository. It
does not guarantee "wiping" the dashboard in the same way, but golang/ci
bucket builds will always replace older builds for the same commit on
the LUCI dashboards (and our dashboard too).

Fixes golang/go#36131.
Fixes golang/go#27055 by making it obsolete.

Change-Id: I5646982d285145622d9951a7bebbcd9025caf47f
Reviewed-on: https://go-review.googlesource.com/c/build/+/600855
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>