cmd/gopherbot: fix repository reference in "close luci-config issues"
I anticipated that the commit link would need to be absolute, since
commits to the luci-config branch get mirrored to a different GitHub
repo than where the issue is being closed, but accidentally pointed it
to the wrong repository. It should be where the repo where the commit
is, not where the referenced issue is.

Change-Id: Ieba4c3146355b4f6c66ef648091905168603cd8d
Reviewed-on: https://go-review.googlesource.com/c/build/+/601155
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>