main.star: mark darwin-amd64 host as slower
We're aware of an issue where the current darwin-amd64 hosts are slower
than we expect, primarily in the context of the darwin-amd64-longtest
builders. Various cmd/go tests often end up failing with a timeout.

As a step in investigating this, mark it as slow so that the longtest
builders have a longer timeout and we can get more information out of
these builders.

For golang/go#65040.

Change-Id: I18fa6fc2ef944d40ad4ad1a96e572d674023f279
Reviewed-on: https://go-review.googlesource.com/c/build/+/600998
Reviewed-by: Cherry Mui <cherryyz@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>