cmd/watchflakes: coalesce a package-level failure to a test failure in the same package
Recently we added reporting package-level failures (that is not
associated to a specific test) to LUCI. In some cases, e.g. a test
timeout could also cause a package-level failure. E.g.
https://ci.chromium.org/b/8745232168515795905 . Currently
watchflakes reports them as two issues. But they are the same
cause. This CL changes it to combine them, when there is a
package-level failure and also a test failure in the same package.

Also update splitTestID to handle package-level failure. If it
cannot split, treat the whole TestID as package, not the test.

Fixes golang/go#68597.

Change-Id: I0679a0345194589e84b351ea7519b2d4da2731e3
Reviewed-on: https://go-review.googlesource.com/c/build/+/601436
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>