cmd/gomote: emit repro test commands with CWD set to package directory
Currently the test commands emitted by the repro subcommand set CWD to
the repo's top-level directory and specify a relative package path for
the package to test. This has a problem in that the test command is
incorrect for reproducing tests from nested Go modules. Since the CWD is
very likely in a parent module, any attempt to run a test in a nested
module from the parent module's directory will fail.

Fix this with a simple trick: set CWD to the package we want to test.
Then, we're always guaranteed to be executing the test from the right
location, within the submodule.

Thanks to dmitshur@golang.org for pointing this out to me!

Change-Id: I121a39f04d8c0e874a339f88d43c01237118bc40
Reviewed-on: https://go-review.googlesource.com/c/build/+/601795
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>