cmd/watchflakes: post new issues to GitHub at the end
Currently, if -post is set, when we encounter a new issue, it
immediately creates an issue on GitHub. This CL changes it to post
at end. This gives us a chance to modify the new issue before
posting. This also makes posting actions together.

Change-Id: I8252b0fde0373134378e61c440b5b62c3cdc514d
Reviewed-on: https://go-review.googlesource.com/c/build/+/602375
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>