main.star: add example, xerrors, and scratch projects
These previously had presubmit, so add them back with presubmit. To
retain the uniformity of our existing policy, we also let them get added
as postsubmit.

We add them as PT.TOOL so that they only run on a small subset of
platforms, but they're probably tested with *way* more configurations
than they need to be. This is generally harmless, though if it becomes a
problem we can prune the builders more aggressively.

Change-Id: I9a8317dd8473e4efa7c917b9b7aa25a089034bb2
Reviewed-on: https://go-review.googlesource.com/c/build/+/608495
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Pratt <mpratt@google.com>