internal/task: don't introduce toolchain directive in tagging workflow
The golang.org/x repo tagging/updating workflow will inadvertently add
a 'toolchain go1.23.0' directive in a few modules in its next upcoming
run. This is because we use the latest Go toolchain to generate go.mod
file changes, and default behavior of the go command:

> For repeatability, any command that updates the go line also updates
> the toolchain line to record its own toolchain name.

However, right now we want to keep the toolchain directive at go1.22.0
or so, to avoid module authors from needing to remove it if trying to
use an older (but still supported) version of the Go toolchain. Also
see the rationale at https://go.dev/design/69095-x-repo-continuous-go#why-not-bump-the-toolchain-lines-too.

Modify the tagging/updating workflow to check if a toolchain directive
wasn't already there (the common case at this time), and drop it at the
end if so.

For golang/go#48523.

Change-Id: I9f2fb5b02e61fcdeed0ba43bbcc9fb69a1f1efb9
Reviewed-on: https://go-review.googlesource.com/c/build/+/609396
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Pratt <mpratt@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>