internal/task: drop tagx:compat magic comment in tagging workflow
It's no longer used, and not expected to be needed in the future.
As of Go 1.21, the go directive is a more direct way of stating the
minimum Go version to stay compatible with.

For golang/go#48523.

Change-Id: I3932a103982a6843194195c99304bf265217b8d1
Reviewed-on: https://go-review.googlesource.com/c/build/+/609397
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Pratt <mpratt@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>