internal/releasetargets: drop 1.21 targets, add 1.24 targets
The Go 1.21 entries are no longer needed now that go1.23.0 is out.

The list for tip is the same as for Go 1.23, except that the windows/arm
port is not included because it's currently marked broken (CL 601777).

I considered renaming allReleases to something more appropriate since
it tracks only first class ports now, but leaving that to a future CL.

For golang/go#40561.
For golang/go#68552.

Change-Id: Ic61e797a125aef5b9cc2782b87ea558e7c88035c
Reviewed-on: https://go-review.googlesource.com/c/build/+/612555
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Tim King <taking@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>