internal/gomote: specify workdir to buildlet as an absolute path
Currently, the workdir passed to the buildlet is a relative path, but
this breaks an assumption in the buildlet about the --workdir flag
always being absolute or empty (if empty, it is set to an absolute
path). This results in a bunch of things breaking because a bunch of
codepaths in handleExec in the buildlet expect an absolute path.

At least, that's my current theory. If this doesn't work, then I'll
revert the change to set --workdir and dig deeper.

Change-Id: I3a484e336f997804c0aaaaae02d0af0f00cb6e9b
Reviewed-on: https://go-review.googlesource.com/c/build/+/616917
Reviewed-by: Michael Pratt <mpratt@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>