env/freebsd-amd64: exit daemon when bootstrapswarm has a zero exit code
Part of the contract of running swarming_bot is that we need to keep
restarting it if it has a non-zero exit code. But if it has a zero exit
code, we must NOT restart it. Currently the daemon just always keeps
restarting it.

Do this by writing the pid of the daemon supervisor and using that to
kill the supervisor when we get a zero exit code from bootstrapswarm.
bootstrapswarm will exit with a zero exit code only if swarming_bot
does.

For golang/go#61095.

Change-Id: I92bdea3fa35fed0883cd9e996aa231e26cdb7155
Reviewed-on: https://go-review.googlesource.com/c/build/+/622040
Auto-Submit: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>