internal/relui: serve 404 when workflow doesn't exist, fix DB tests
Relui is expected to serve 404 when a workflow that doesn't exist is
requested. Its tests, ones that run only when a database is available,
also expect that. Update a few places after CL 593915 accordingly to
serve 404 instead of 500, and update TestServerStopWorkflow to create
a workflow in the DB, since stopWorkflowHandler now fetches it as part
of checking ACLs.

For golang/go#68114.

Change-Id: I90f12a431b1f97d8be33d6404eb7e2064e50f688
Reviewed-on: https://go-review.googlesource.com/c/build/+/624076
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>