perf: fork and use benchtab over benchseries
benchseries seems to fall over on some of our benchmarks and it's
difficult to say why. But, benchstat works just fine.

This CL forks the internals of benchstat and rewrites benchCompare to
use them.

The comparison made by benchstat is slightly different than the one
made by benchseries, but it's not clear to me how. It might make sense
to try and flush InfluxDB and refill it with the new comparison, but I
suspect it's probably not worth it.

Change-Id: I7e1ea887106e6d7bd9bb310b6ee34cef47efde19
Reviewed-on: https://go-review.googlesource.com/c/build/+/623275
Reviewed-by: Michael Pratt <mpratt@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>