perf: fork and use benchtab over benchseries
benchseries seems to fall over on some of our benchmarks and it's
difficult to say why. But, benchstat works just fine.

This CL forks the internals of benchstat and rewrites benchCompare to
use them.

The comparison made by benchstat is slightly different than the one
made by benchseries, but it's not clear to me how. It might make sense
to try and flush InfluxDB and refill it with the new comparison, but I
suspect it's probably not worth it.

This is a reapplication of go.dev/cl/623275 to correctly propagate
bootstrapped ratio data instead of just a summary of the direct
benchmark data.

Change-Id: I24d161130934e59aee4082629e8d0698fef3fde5
Reviewed-on: https://go-review.googlesource.com/c/build/+/626640
Reviewed-by: Michael Pratt <mpratt@google.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>