cmd/watchflakes: do not exit in the case of a mismatch
When watchflakes encountered a build with a mismatch between
a builder definition and the input revision it would log the error and
exit. This change changes the behavior to log that the mismatch exists
and skip the build. This was causing watchflakes to continuously exit
and never process builds past the point of the encountered mismatch.

Fixes golang/go#70091

Change-Id: Ic3e2957a8e501f6dfa8bbbc458db40f925a0877d
Reviewed-on: https://go-review.googlesource.com/c/build/+/623995
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Carlos Amedee <carlos@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>