internal/workflow: run at most one expansion at a time
We've inadvertently been running expansions concurrently during minor
releases. For better or worse, it happened to be seemingly fine, and
problems started to be noticeable only when needing to restart or
approve tasks, which failed with puzzling "unknown task" errors.

It's useful to be able to plan builders for the two Go releases in
parallel, and it's completely fine for it not to happen concurrently.
Instead of getting the workflow to arrange for that, it seems we can
do it in the workflow package itself.

The new TestManualRetryMultipleExpansions test fails before the change,
and passes after.

Fixes golang/go#70249.

Change-Id: Id87323f77f573d9ac364010dfc0b8581e57ce9b8
Reviewed-on: https://go-review.googlesource.com/c/build/+/626335
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>